Wed, Jan 10, 2024, 4:45 PM +08:00

+++ filter-less/helpers.c
@@ -1,4 +1,5 @@
 #include "helpers.h"
+#include <math.h>
 
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
@@ -24,7 +25,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
 // Convert image to sepia
 void sepia(int height, int width, RGBTRIPLE image[height][width])
 {
-    int sepiaRed, sepiaGreen, sepiaBlue;
+    float sepiaRed, sepiaGreen, sepiaBlue;
     int max = 255;
     // Loop through all pixels
     for (int i = 0; i < height; i++)
@@ -49,7 +50,7 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
 }
 
 // Return which is lower
-int min(int *a, int *b)
+float min(float *a, float *b)
 {
     return (*a < *b) ? *a : *b;
 }
