+++ practice/week4/practice.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include <math.h>
 
-int min(int *a, int *b);
+float min(float *a, float *b);
 
 int main(void)
 {
@@ -14,7 +15,7 @@ int main(void)
     printf("%i %i %i\n", rgbtRed, rgbtGreen, rgbtBlue);
 }
 
-float min(int *a, int *b)
+float min(float *a, float *b)
 {
     return (*a < *b) ? *a : *b;
 }
